+++ example/react_icu_withHOC/src/i18n.js
@@ -3,13 +3,15 @@ import Backend from 'i18next-xhr-backend';
 import LanguageDetector from 'i18next-browser-languagedetector';
 import { reactI18nextModule } from 'react-i18next';
 
-import ICU from 'i18next-icu'
+import ICU from 'i18next-icu';
 import de from 'i18next-icu/locale-data/de'; // or dynamically like: https://github.com/locize/locize-react-intl-example/blob/master/src/locize/index.js#L53
 
 i18n
-  .use(new ICU({
-    localeData: de // you also can pass in array of localeData
-  }))
+  .use(
+    new ICU({
+      localeData: de, // you also can pass in array of localeData
+    })
+  )
   .use(Backend)
   .use(LanguageDetector)
   .use(reactI18nextModule)
@@ -20,6 +22,9 @@ i18n
     ns: ['translations'],
     defaultNS: 'translations',
 
+    nsSeparator: false,
+    keySeparator: false,
+
     debug: true,
 
     interpolation: {
@@ -27,9 +32,8 @@ i18n
     },
 
     react: {
-      wait: true
-    }
+      wait: true,
+    },
   });
-
 
 export default i18n;