+++ rollup.config.js
@@ -1,34 +1,32 @@
 import babel from 'rollup-plugin-babel';
-import uglify from 'rollup-plugin-uglify';
 import commonjs from 'rollup-plugin-commonjs';
 import nodeResolve from 'rollup-plugin-node-resolve';
+import { terser } from 'rollup-plugin-terser';
 import { argv } from 'yargs';
 
 const format = argv.format || argv.f || 'iife';
 const compress = argv.uglify;
 
 const babelOptions = {
-  exclude: 'node_modules/**',
-  presets: ['es2015-rollup', 'stage-0'],
-  babelrc: false
+  // babelrc: false,
 };
 
-const dest = {
+const file = {
   amd: `dist/amd/react-i18next${compress ? '.min' : ''}.js`,
   umd: `dist/umd/react-i18next${compress ? '.min' : ''}.js`,
-  iife: `dist/iife/react-i18next${compress ? '.min' : ''}.js`
+  iife: `dist/iife/react-i18next${compress ? '.min' : ''}.js`,
 }[format];
 
 export default {
-  entry: 'src/index.js',
-  format,
+  input: 'src/index.js',
+  plugins: [babel(babelOptions), nodeResolve({ jsnext: true, main: true }), commonjs({})].concat(
+    compress ? terser() : []
+  ),
   external: ['react', 'react-dom', 'prop-types'],
-  plugins: [
-    babel(babelOptions),
-    nodeResolve({ jsnext: true }),
-    commonjs()
-  ].concat(compress ? uglify() : []),
-  moduleName: 'reactI18next',
-  moduleId: 'reactI18next',
-  dest
+  // moduleId: 'react-i18next',
+  output: {
+    name: 'react-i18next',
+    format,
+    file,
+  },
 };