+++ src/context.js
@@ -12,7 +12,7 @@ let defaultOptions = {
   omitBoundRerender: true,
 };
 
-let i18n;
+let i18nInstance;
 
 export function setDefaults(options) {
   defaultOptions = { ...defaultOptions, ...options };
@@ -23,11 +23,11 @@ export function getDefaults() {
 }
 
 export function setI18n(instance) {
-  i18n = instance;
+  i18nInstance = instance;
 }
 
 export function getI18n() {
-  return i18n;
+  return i18nInstance;
 }
 
 export const reactI18nextModule = {
@@ -42,7 +42,7 @@ export const reactI18nextModule = {
 export const I18nContext = createReactContext();
 
 // hoc for context
-export function withContext() {
+export function withI18n() {
   return function Wrapper(WrappedComponent) {
     class WithContext extends Component {
       constructor(props, context) {
@@ -56,17 +56,43 @@ export function withContext() {
       }
 
       render() {
+        // extra props like ref
         const extraProps = {
           ref: c => {
             this.wrapped = c;
           },
         };
 
+        let i18nOptions = this.props.i18nOptions;
+
+        // as default we add i18n, basic t function and i18nOptions from setI18n
+        // those get overridden by values passed by I18nContext.Provider <- eg. set in I18nextProvider
+        const i18n = this.props.i18n || getI18n();
+
+        if (!i18nOptions) {
+          const possibleI18nOptionsFromProps = Object.keys(defaultOptions).reduce((mem, k) => {
+            if (this.props[k]) mem[k] = this.props[k];
+            return mem;
+          }, {});
+          i18nOptions = {
+            ...getDefaults(),
+            ...(i18n && i18n.options && i18n.options.react),
+            ...possibleI18nOptionsFromProps,
+          };
+        }
+
+        if (i18n) {
+          extraProps.i18n = i18n;
+          extraProps.t = i18n.t.bind(i18n);
+          extraProps.lng = i18n.language;
+          extraProps.i18nOptions = i18nOptions;
+        }
+
         return React.createElement(I18nContext.Consumer, null, ctx =>
           React.createElement(WrappedComponent, {
+            ...extraProps,
             ...ctx,
             ...this.props,
-            ...extraProps,
           })
         );
       }