+++ test/NamespacesConsumer.spec.js
@@ -1,11 +1,13 @@
-import { I18n } from '../src/I18n';
-jest.unmock('../src/I18n');
+import { NamespacesConsumerComponent } from '../src/NamespacesConsumer';
+
+jest.unmock('../src/NamespacesConsumer');
 
 const i18n = {
   language: 'en',
   languages: ['en'],
   options: {
     defaultNS: 'defaultNS',
+    nsMode: 'fallback',
   },
   services: {
     resourceStore: {
@@ -18,14 +20,20 @@ const i18n = {
   loadNamespaces: () => {},
 };
 
-describe('I18n', () => {
+describe('NamespacesConsumer', () => {
   it('sets initial ready state to false if resource bundles are not loaded', () => {
-    const notLoadedInstance = new I18n({ i18n, ns: ['notLoadedNS', 'alreadyLoadedNS'] }, {});
+    const notLoadedInstance = new NamespacesConsumerComponent(
+      { i18n, ns: ['notLoadedNS', 'alreadyLoadedNS'], i18nOptions: { ...i18n.options } },
+      {}
+    );
     expect(notLoadedInstance.state.ready).toBe(false);
   });
 
   it('sets initial ready state to true if resource bundles are already loaded', () => {
-    const notLoadedInstance = new I18n({ i18n, ns: ['alreadyLoadedNS'] }, {});
+    const notLoadedInstance = new NamespacesConsumerComponent(
+      { i18n, ns: ['alreadyLoadedNS'], i18nOptions: { ...i18n.options } },
+      {}
+    );
     expect(notLoadedInstance.state.ready).toBe(true);
   });
 });