+++ test/i18nextProvider.spec.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import I18nextProvider from '../src/I18nextProvider';
+import { I18nextProvider } from '../src/I18nextProvider';
 import { I18nContext } from '../src/context';
 
 jest.unmock('../src/I18nextProvider');