+++ test/trans.spec.js
@@ -1,13 +1,11 @@
-jest.unmock('../src/Trans');
 import React from 'react';
 import PropTypes from 'prop-types';
-import { Trans } from '../src/Trans';
+import { TransComponent } from '../src/Trans';
+jest.unmock('../src/Trans');
 
 describe('trans', () => {
   it('should have some stuff', () => {
-    expect(Trans.propTypes.i18n)
-      .toBe(PropTypes.object);
-    expect(Trans.propTypes.t)
-      .toBe(PropTypes.func);
+    expect(TransComponent.propTypes.i18n).toBe(PropTypes.object);
+    expect(TransComponent.propTypes.t).toBe(PropTypes.func);
   });
 });