+++ test/withNamespaces.render.wait.spec.js
@@ -2,31 +2,24 @@ import React from 'react';
 import { shallow, render, mount } from 'enzyme';
 import i18n from './i18n';
 import BackendMock from './backendMock';
-import translate from '../src/translate';
+import { withNamespaces } from '../src/withNamespaces';
 
 const newI18n = i18n.createInstance();
 const backend = new BackendMock();
-newI18n
-  .use(backend)
-  .init({
-    lng: 'en',
-    fallbackLng: 'en',
+newI18n.use(backend).init({
+  lng: 'en',
+  fallbackLng: 'en',
 
-    interpolation: {
-      escapeValue: false, // not needed for react!!
-    }
-  })
+  interpolation: {
+    escapeValue: false, // not needed for react!!
+  },
+});
 
-
-describe('translate wait', () => {
-  const TestElement = ({ t }) => {
-    return (
-      <div>{t('key1')}</div>
-    );
-  }
+describe('withNamespaces wait', () => {
+  const TestElement = ({ t }) => <div>{t('key1')}</div>;
 
   it('should wait for correct translation', () => {
-    const HocElement = translate(['common'], { wait: true })(TestElement);
+    const HocElement = withNamespaces(['common'], { wait: true })(TestElement);
 
     const wrapper = mount(<HocElement i18n={i18n} />);
     // console.log(wrapper.debug());
@@ -40,12 +33,13 @@ describe('translate wait', () => {
     }, 50);
   });
 
-
   it('should global wait for correct translation', () => {
     // reinit global
-    newI18n.init({ react: { wait: true }});
+    newI18n.init({ react: { wait: true } });
 
-    const HocElement = translate(['common'], { /* not use this -> wait: true */ })(TestElement);
+    const HocElement = withNamespaces(['common'], {
+      /* not use this -> wait: true */
+    })(TestElement);
 
     const wrapper = mount(<HocElement i18n={newI18n} />);
     // console.log(wrapper.debug());